requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.21.0
clip-by-openai>=1.0
faiss-cpu>=1.7.0
numpy>=1.21.0
tqdm>=4.64.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
Pillow>=8.3.0